Moved around property code to avoid circular dependency
