+++ sympl/__init__.py
@@ -14,13 +14,14 @@ from ._core.dataarray import DataArray
 from ._core.constants import (
     get_constant, set_constant, set_condensible_name, reset_constants,
     get_constants_string)
-from ._core.tracers import register_tracer, get_tracer_unit_dict, get_tracer_names
+from ._core.tracers import (
+    register_tracer, get_tracer_unit_dict, get_tracer_input_properties, get_tracer_names)
 from ._core.util import (
     ensure_no_shared_keys,
     get_numpy_array, jit,
     restore_dimensions,
-    get_component_aliases,
-    combine_component_properties)
+    get_component_aliases)
+from sympl._core.combine_properties import combine_component_properties
 from ._core.units import units_are_same, units_are_compatible, is_valid_unit
 from sympl._core.get_np_arrays import get_numpy_arrays_with_properties
 from sympl._core.restore_dataarray import restore_data_arrays_with_properties
@@ -47,7 +48,7 @@ __all__ = (
     get_constants_string, TimeDifferencingWrapper,
     ensure_no_shared_keys,
     get_numpy_array, jit,
-    register_tracer, get_tracer_unit_dict, get_tracer_names,
+    register_tracer, get_tracer_unit_dict, get_tracer_input_properties, get_tracer_names,
     restore_dimensions, get_numpy_arrays_with_properties,
     restore_data_arrays_with_properties,
     initialize_numpy_arrays_with_properties,
